🗺️ Atlas: [architectural change] Break circular dependencies in autumn-web and fix coverage
🕸️ Tangle: Two circular dependency cycles existed:
1. `db` -> `state` -> `db` (caused by `Db` extractor tightly coupling to `AppState`).
2. `state` -> `actuator` -> `state` (caused by `actuator` holding telemetry types that `AppState` used).

📐 Blueprint:
1. Extracted telemetry and diagnostic types (`LogLevels`, `TaskRegistry`, etc.) from `actuator.rs` into a new `diagnostics.rs` module. Both `state.rs` and `actuator.rs` now depend on `diagnostics.rs`.
2. Created a `DbState` trait in `db.rs` and made the `Db` extractor generic over `S: DbState + Send + Sync`. `AppState` now implements this trait, completely removing `db`'s dependency on `state`.
3. Moved relevant unit tests to `diagnostics.rs` to ensure tests run properly and coverage targets are met. Fixed unclosed brace in tests preventing compilation. Moved test items back to avoid clippy warnings.

🧱 Stability: Enforced strict unidirectional dependencies and eliminated the `AppState` "God Struct" coupling from sub-modules. Reduced coupling.

🔬 Verification: Builds successfully (`cargo check`), strict separation verified with a circular dependency detection script, and all tests pass. Cleaned up scratch files.
