+++ autumn/src/actuator.rs
@@ -8,12 +8,375 @@
 //! - **prod**: only health, info, and metrics
 
 use std::collections::HashMap;
+use std::sync::{Arc, RwLock};
 
 use axum::Json;
 use axum::extract::{Path, State};
 use axum::http::StatusCode;
 use axum::response::IntoResponse;
 use serde::{Deserialize, Serialize};
+
+use crate::state::AppState;
+
+// ── Shared types for AppState ──────────────────────────────────
+
+/// Runtime log level management for the `/actuator/loggers` endpoint.
+///
+/// Stores the current effective log level and per-logger overrides.
+/// Changes are ephemeral -- they reset on restart.
+#[derive(Clone)]
+pub struct LogLevels {
+    inner: Arc<RwLock<LogLevelsInner>>,
+}
+
+struct LogLevelsInner {
+    /// The current global log level.
+    current_level: String,
+    /// Per-logger level overrides applied at runtime.
+    logger_overrides: HashMap<String, String>,
+}
+
+impl LogLevels {
+    /// Create a new `LogLevels` with the given initial level.
+    #[must_use]
+    pub fn new(initial_level: &str) -> Self {
+        Self {
+            inner: Arc::new(RwLock::new(LogLevelsInner {
+                current_level: initial_level.to_string(),
+                logger_overrides: HashMap::new(),
+            })),
+        }
+    }
+
+    /// Get the current global log level.
+    #[must_use]
+    pub fn current_level(&self) -> String {
+        self.inner
+            .read()
+            .map_or_else(|_| "info".to_string(), |guard| guard.current_level.clone())
+    }
+
+    /// Get all per-logger overrides.
+    #[must_use]
+    pub fn logger_overrides(&self) -> HashMap<String, String> {
+        self.inner
+            .read()
+            .map(|guard| guard.logger_overrides.clone())
+            .unwrap_or_default()
+    }
+
+    /// Set the level for a specific logger. Returns the previous level if any.
+    #[must_use]
+    pub fn set_logger_level(&self, name: &str, level: &str) -> Option<String> {
+        if let Ok(mut guard) = self.inner.write() {
+            // Prevent unbounded memory growth from arbitrary logger names
+            if guard.logger_overrides.len() >= 1000 && !guard.logger_overrides.contains_key(name) {
+                return None;
+            }
+
+            let previous = guard.logger_overrides.get(name).cloned();
+            guard
+                .logger_overrides
+                .insert(name.to_string(), level.to_string());
+            // If setting the root level, update current_level too
+            if name == "root" || name.is_empty() {
+                let prev = Some(guard.current_level.clone());
+                guard.current_level = level.to_string();
+                return prev;
+            }
+            previous
+        } else {
+            None
+        }
+    }
+}
+
+impl std::fmt::Debug for LogLevels {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        f.debug_struct("LogLevels")
+            .field("current_level", &self.current_level())
+            .finish()
+    }
+}
+
+/// Scheduled task status information.
+#[derive(Debug, Clone, Serialize)]
+pub struct TaskStatus {
+    /// The schedule description (e.g., "every 5m" or "cron 0 0 * * *").
+    pub schedule: String,
+    /// Current task state.
+    pub status: String,
+    /// Last time the task ran (ISO 8601), if ever.
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub last_run: Option<String>,
+    /// Duration of last run in milliseconds.
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub last_duration_ms: Option<u64>,
+    /// Result of last run.
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub last_result: Option<String>,
+    /// Last error message, if the task failed.
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub last_error: Option<String>,
+    /// Total number of times the task has run.
+    pub total_runs: u64,
+    /// Total number of failures.
+    pub total_failures: u64,
+}
+
+/// Registry of scheduled tasks and their runtime status.
+#[derive(Clone)]
+pub struct TaskRegistry {
+    inner: Arc<RwLock<HashMap<String, TaskStatus>>>,
+}
+
+impl TaskRegistry {
+    /// Create a new empty task registry.
+    #[must_use]
+    pub fn new() -> Self {
+        Self {
+            inner: Arc::new(RwLock::new(HashMap::new())),
+        }
+    }
+
+    /// Register a task with its schedule description.
+    pub fn register(&self, name: &str, schedule: &str) {
+        if let Ok(mut guard) = self.inner.write() {
+            guard.insert(
+                name.to_string(),
+                TaskStatus {
+                    schedule: schedule.to_string(),
+                    status: "idle".to_string(),
+                    last_run: None,
+                    last_duration_ms: None,
+                    last_result: None,
+                    last_error: None,
+                    total_runs: 0,
+                    total_failures: 0,
+                },
+            );
+        }
+    }
+
+    /// Record that a task started running.
+    pub fn record_start(&self, name: &str) {
+        if let Ok(mut guard) = self.inner.write() {
+            if let Some(task) = guard.get_mut(name) {
+                task.status = "running".to_string();
+            }
+        }
+    }
+
+    /// Record that a task completed successfully.
+    pub fn record_success(&self, name: &str, duration_ms: u64) {
+        if let Ok(mut guard) = self.inner.write() {
+            if let Some(task) = guard.get_mut(name) {
+                task.status = "idle".to_string();
+                task.last_run = Some(chrono::Utc::now().to_rfc3339());
+                task.last_duration_ms = Some(duration_ms);
+                task.last_result = Some("ok".to_string());
+                task.last_error = None;
+                task.total_runs += 1;
+            }
+        }
+    }
+
+    /// Record that a task failed.
+    pub fn record_failure(&self, name: &str, duration_ms: u64, error: &str) {
+        if let Ok(mut guard) = self.inner.write() {
+            if let Some(task) = guard.get_mut(name) {
+                task.status = "idle".to_string();
+                task.last_run = Some(chrono::Utc::now().to_rfc3339());
+                task.last_duration_ms = Some(duration_ms);
+                task.last_result = Some("failed".to_string());
+                task.last_error = Some(error.to_string());
+                task.total_runs += 1;
+                task.total_failures += 1;
+            }
+        }
+    }
+
+    /// Get a snapshot of all task statuses.
+    #[must_use]
+    pub fn snapshot(&self) -> HashMap<String, TaskStatus> {
+        self.inner
+            .read()
+            .map(|guard| guard.clone())
+            .unwrap_or_default()
+    }
+}
+
+impl Default for TaskRegistry {
+    fn default() -> Self {
+        Self::new()
+    }
+}
+
+impl std::fmt::Debug for TaskRegistry {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        f.debug_struct("TaskRegistry")
+            .field("count", &self.snapshot().len())
+            .finish()
+    }
+}
+
+/// Resolved config property with source provenance.
+#[derive(Debug, Clone, Serialize, Deserialize)]
+pub struct ConfigProperty {
+    /// The resolved value (redacted if sensitive).
+    pub value: serde_json::Value,
+    /// Where the value came from.
+    pub source: String,
+}
+
+/// Collection of resolved config properties with source tracking.
+#[derive(Debug, Clone, Default)]
+pub struct ConfigProperties {
+    inner: Arc<RwLock<HashMap<String, ConfigProperty>>>,
+}
+
+impl ConfigProperties {
+    /// Build config properties with source tracking from the loaded config.
+    #[must_use]
+    pub fn from_config(config: &crate::config::AutumnConfig) -> Self {
+        let profile = config.profile.as_deref().unwrap_or("default");
+        let defaults = crate::config::AutumnConfig::default();
+
+        let mut props = HashMap::new();
+        let profile_str = profile.to_string();
+
+        // Server properties
+        Self::track_property(
+            &mut props,
+            "server.host",
+            &config.server.host,
+            &defaults.server.host,
+            &profile_str,
+        );
+        Self::track_property(
+            &mut props,
+            "server.port",
+            &config.server.port.to_string(),
+            &defaults.server.port.to_string(),
+            &profile_str,
+        );
+        Self::track_property(
+            &mut props,
+            "server.shutdown_timeout_secs",
+            &config.server.shutdown_timeout_secs.to_string(),
+            &defaults.server.shutdown_timeout_secs.to_string(),
+            &profile_str,
+        );
+
+        // Database properties
+        let db_url = config.database.url.as_deref().unwrap_or("").to_string();
+        Self::track_property(&mut props, "database.url", &db_url, "", &profile_str);
+        Self::track_property(
+            &mut props,
+            "database.pool_size",
+            &config.database.pool_size.to_string(),
+            &defaults.database.pool_size.to_string(),
+            &profile_str,
+        );
+
+        // Log properties
+        Self::track_property(
+            &mut props,
+            "log.level",
+            &config.log.level,
+            &defaults.log.level,
+            &profile_str,
+        );
+        Self::track_property(
+            &mut props,
+            "log.format",
+            &format!("{:?}", config.log.format),
+            &format!("{:?}", defaults.log.format),
+            &profile_str,
+        );
+
+        // Health properties
+        Self::track_property(
+            &mut props,
+            "health.path",
+            &config.health.path,
+            &defaults.health.path,
+            &profile_str,
+        );
+        Self::track_property(
+            &mut props,
+            "health.detailed",
+            &config.health.detailed.to_string(),
+            &defaults.health.detailed.to_string(),
+            &profile_str,
+        );
+
+        // Actuator properties
+        Self::track_property(
+            &mut props,
+            "actuator.prefix",
+            &config.actuator.prefix,
+            &defaults.actuator.prefix,
+            &profile_str,
+        );
+        Self::track_property(
+            &mut props,
+            "actuator.sensitive",
+            &config.actuator.sensitive.to_string(),
+            &defaults.actuator.sensitive.to_string(),
+            &profile_str,
+        );
+
+        Self {
+            inner: Arc::new(RwLock::new(props)),
+        }
+    }
+
+    /// Track a single config property, determining its source by checking
+    /// for env var overrides and comparing against defaults.
+    fn track_property(
+        props: &mut HashMap<String, ConfigProperty>,
+        key: &str,
+        value: &str,
+        default_value: &str,
+        profile: &str,
+    ) {
+        // Check if there's an env var override
+        let env_key = format!("AUTUMN_{}", key.replace('.', "__").to_uppercase());
+        let source = if std::env::var(&env_key).is_ok() {
+            env_key
+        } else if value != default_value && (profile == "dev" || profile == "prod") {
+            format!("profile_default:{profile}")
+        } else if value != default_value {
+            "autumn.toml".to_string()
+        } else {
+            "default".to_string()
+        };
+
+        let display_value = if should_redact(key) {
+            serde_json::Value::String("****".into())
+        } else {
+            serde_json::Value::String(value.to_string())
+        };
+
+        props.insert(
+            key.to_string(),
+            ConfigProperty {
+                value: display_value,
+                source,
+            },
+        );
+    }
+
+    /// Get a snapshot of all properties.
+    #[must_use]
+    pub fn snapshot(&self) -> HashMap<String, ConfigProperty> {
+        self.inner
+            .read()
+            .map(|guard| guard.clone())
+            .unwrap_or_default()
+    }
+}
 
 // ── Health ──────────────────────────────────────────────────────
 
@@ -44,7 +407,7 @@ struct DatabaseCheck {
 
 /// `GET /actuator/health`
 #[allow(unused_variables, clippy::useless_let_if_seq)]
-pub async fn health(State(state): State<crate::state::AppState>) -> impl IntoResponse {
+pub async fn health(State(state): State<AppState>) -> impl IntoResponse {
     let db_check;
     let overall_healthy;
 
@@ -123,7 +486,7 @@ struct RuntimeInfo {
 }
 
 /// `GET /actuator/info`
-pub(crate) async fn info(State(state): State<crate::state::AppState>) -> Json<ActuatorInfo> {
+pub(crate) async fn info(State(state): State<AppState>) -> Json<ActuatorInfo> {
     Json(ActuatorInfo {
         app: AppInfo {
             name: std::env::var("CARGO_PKG_NAME").unwrap_or_else(|_| "unknown".into()),
@@ -148,8 +511,24 @@ pub(crate) struct ActuatorEnv {
     properties: std::collections::HashMap<String, serde_json::Value>,
 }
 
+/// Keys that trigger value redaction.
+const REDACT_PATTERNS: &[&str] = &[
+    "password",
+    "secret",
+    "key",
+    "token",
+    "credential",
+    "auth",
+    "url",
+];
+
+fn should_redact(key: &str) -> bool {
+    let lower = key.to_lowercase();
+    REDACT_PATTERNS.iter().any(|p| lower.contains(p))
+}
+
 /// `GET /actuator/env` — only available when actuator sensitive mode is enabled.
-pub(crate) async fn env_endpoint(State(state): State<crate::state::AppState>) -> Json<ActuatorEnv> {
+pub(crate) async fn env_endpoint(State(state): State<AppState>) -> Json<ActuatorEnv> {
     let props = vec![
         ("server.host", "127.0.0.1"),
         ("server.port", "3000"),
@@ -162,7 +541,7 @@ pub(crate) async fn env_endpoint(State(state): State<crate::state::AppState>) ->
     let redacted: std::collections::HashMap<String, serde_json::Value> = props
         .into_iter()
         .map(|(k, v)| {
-            let val = if crate::diagnostics::should_redact(k) {
+            let val = if should_redact(k) {
                 serde_json::Value::String("****".into())
             } else {
                 serde_json::Value::String(v.into())
@@ -181,10 +560,8 @@ pub(crate) async fn env_endpoint(State(state): State<crate::state::AppState>) ->
 
 /// `GET /actuator/metrics` -- request metrics, latency, status codes, DB pool stats.
 #[allow(unused_variables, unused_mut)]
-pub(crate) async fn metrics_endpoint(
-    State(state): State<crate::state::AppState>,
-) -> Json<serde_json::Value> {
-    let snapshot = state.metrics().snapshot();
+pub(crate) async fn metrics_endpoint(State(state): State<AppState>) -> Json<serde_json::Value> {
+    let snapshot = state.metrics.snapshot();
     let mut result = serde_json::to_value(&snapshot).unwrap_or_default();
 
     // Include DB pool stats if available
@@ -207,10 +584,8 @@ pub(crate) async fn metrics_endpoint(
 // ── Config Properties (sensitive) ──────────────────────────────
 
 /// `GET /actuator/configprops` -- all config properties with source tracking.
-pub(crate) async fn configprops_endpoint(
-    State(state): State<crate::state::AppState>,
-) -> Json<serde_json::Value> {
-    let props = state.config_props().snapshot();
+pub(crate) async fn configprops_endpoint(State(state): State<AppState>) -> Json<serde_json::Value> {
+    let props = state.config_props.snapshot();
 
     Json(serde_json::json!({
         "active_profile": state.profile(),
@@ -232,13 +607,11 @@ pub(crate) struct LoggersResponse {
 }
 
 /// `GET /actuator/loggers` -- view current log levels.
-pub(crate) async fn loggers_get(
-    State(state): State<crate::state::AppState>,
-) -> Json<LoggersResponse> {
+pub(crate) async fn loggers_get(State(state): State<AppState>) -> Json<LoggersResponse> {
     Json(LoggersResponse {
-        current_level: state.log_levels().current_level(),
+        current_level: state.log_levels.current_level(),
         available_levels: AVAILABLE_LEVELS.to_vec(),
-        loggers: state.log_levels().logger_overrides(),
+        loggers: state.log_levels.logger_overrides(),
     })
 }
 
@@ -250,7 +623,7 @@ pub(crate) struct SetLoggerRequest {
 
 /// `PUT /actuator/loggers/{name}` -- change a logger's level at runtime.
 pub(crate) async fn loggers_put(
-    State(state): State<crate::state::AppState>,
+    State(state): State<AppState>,
     Path(name): Path<String>,
     Json(body): Json<SetLoggerRequest>,
 ) -> impl IntoResponse {
@@ -271,7 +644,7 @@ pub(crate) async fn loggers_put(
         );
     }
 
-    let previous = state.log_levels().set_logger_level(&name, &level);
+    let previous = state.log_levels.set_logger_level(&name, &level);
 
     (
         StatusCode::OK,
@@ -286,10 +659,8 @@ pub(crate) async fn loggers_put(
 // ── Tasks (sensitive) ──────────────────────────────────────────
 
 /// `GET /actuator/tasks` -- scheduled task status.
-pub(crate) async fn tasks_endpoint(
-    State(state): State<crate::state::AppState>,
-) -> Json<serde_json::Value> {
-    let tasks = state.task_registry().snapshot();
+pub(crate) async fn tasks_endpoint(State(state): State<AppState>) -> Json<serde_json::Value> {
+    let tasks = state.task_registry.snapshot();
 
     Json(serde_json::json!({
         "scheduled_tasks": tasks,
@@ -301,7 +672,7 @@ pub(crate) async fn tasks_endpoint(
 /// `GET /actuator/tasks/stream` -- stream scheduled task events.
 #[cfg(feature = "ws")]
 pub(crate) async fn tasks_stream_endpoint(
-    State(state): State<crate::state::AppState>,
+    State(state): State<AppState>,
     ws: axum::extract::ws::WebSocketUpgrade,
 ) -> impl IntoResponse {
     ws.on_upgrade(move |mut socket| async move {
@@ -338,7 +709,7 @@ pub(crate) async fn tasks_stream_endpoint(
 ///
 /// In dev mode (or when `actuator.sensitive = true`), all endpoints are
 /// exposed. In prod mode, only health, info, and metrics are available.
-pub fn actuator_router(sensitive: bool) -> axum::Router<crate::state::AppState> {
+pub fn actuator_router(sensitive: bool) -> axum::Router<AppState> {
     let mut router = axum::Router::new()
         .route("/actuator/health", axum::routing::get(health))
         .route("/actuator/info", axum::routing::get(info))
@@ -370,13 +741,12 @@ pub fn actuator_router(sensitive: bool) -> axum::Router<crate::state::AppState> 
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::diagnostics::*;
     use axum::body::Body;
     use axum::http::Request;
     use tower::ServiceExt;
 
-    fn test_state() -> crate::state::AppState {
-        crate::state::AppState {
+    fn test_state() -> AppState {
+        AppState {
             #[cfg(feature = "db")]
             pool: None,
             profile: Some("dev".into()),
@@ -468,15 +838,22 @@ mod tests {
         assert_eq!(resp.status(), StatusCode::NOT_FOUND);
     }
 
-
+    #[test]
+    fn redaction_patterns() {
+        assert!(should_redact("database.url"));
+        assert!(should_redact("api_token"));
+        assert!(should_redact("secret_key"));
+        assert!(!should_redact("server.port"));
+        assert!(!should_redact("log.level"));
+    }
 
     // ── Metrics endpoint tests ─────────────────────────────────
 
     #[tokio::test]
     async fn actuator_metrics_returns_http_stats() {
         let state = test_state();
-        state.metrics().record("GET", "/test", 200, 10);
-        state.metrics().record("POST", "/test", 500, 50);
+        state.metrics.record("GET", "/test", 200, 10);
+        state.metrics.record("POST", "/test", 500, 50);
 
         let app = actuator_router(true).with_state(state);
         let resp = app
@@ -629,7 +1006,7 @@ mod tests {
         assert_eq!(json["status"], "ok");
         assert_eq!(json["message"], "Logger 'autumn_web' set to 'debug'");
 
-        let overrides = state.log_levels().logger_overrides();
+        let overrides = state.log_levels.logger_overrides();
         assert_eq!(
             overrides.get("autumn_web").map(String::as_str),
             Some("debug")
@@ -674,18 +1051,32 @@ mod tests {
         assert_eq!(resp.status(), StatusCode::NOT_FOUND);
     }
 
+    #[test]
+    fn log_levels_set_and_get() {
+        let levels = LogLevels::new("info");
+        assert_eq!(levels.current_level(), "info");
 
+        let _ = levels.set_logger_level("my_crate", "debug");
+        let overrides = levels.logger_overrides();
+        assert_eq!(overrides.get("my_crate").map(String::as_str), Some("debug"));
+    }
 
-
+    #[test]
+    fn log_levels_root_updates_current() {
+        let levels = LogLevels::new("info");
+        let prev = levels.set_logger_level("root", "trace");
+        assert_eq!(prev, Some("info".to_string()));
+        assert_eq!(levels.current_level(), "trace");
+    }
 
     // ── Tasks endpoint tests ───────────────────────────────────
 
     #[tokio::test]
     async fn actuator_tasks_returns_registered_tasks() {
         let state = test_state();
-        state.task_registry().register("cleanup", "every 5m");
-        state.task_registry().record_start("cleanup");
-        state.task_registry().record_success("cleanup", 150);
+        state.task_registry.register("cleanup", "every 5m");
+        state.task_registry.record_start("cleanup");
+        state.task_registry.record_success("cleanup", 150);
 
         let app = actuator_router(true).with_state(state);
         let resp = app
@@ -727,18 +1118,136 @@ mod tests {
         assert_eq!(resp.status(), StatusCode::NOT_FOUND);
     }
 
+    #[test]
+    fn task_registry_records_failure() {
+        let registry = TaskRegistry::new();
+        registry.register("my_task", "cron 0 * * * *");
+        registry.record_start("my_task");
+        registry.record_failure("my_task", 200, "connection refused");
 
+        let snapshot = registry.snapshot();
+        let task = &snapshot["my_task"];
+        assert_eq!(task.status, "idle");
+        assert_eq!(task.total_runs, 1);
+        assert_eq!(task.total_failures, 1);
+        assert_eq!(task.last_result.as_deref(), Some("failed"));
+        assert_eq!(task.last_error.as_deref(), Some("connection refused"));
+    }
 
+    #[test]
+    fn task_registry_empty_snapshot() {
+        let registry = TaskRegistry::new();
+        assert!(registry.snapshot().is_empty());
+    }
+    #[test]
+    fn log_levels_rejects_new_key_at_capacity() {
+        let levels = LogLevels::new("info");
+        // Fill to capacity
+        for i in 0..1000 {
+            let _ = levels.set_logger_level(&format!("logger_{i}"), "debug");
+        }
 
+        // Try to add a new key, should be rejected
+        let result = levels.set_logger_level("logger_1000", "warn");
+        assert_eq!(result, None);
+        assert_eq!(levels.logger_overrides().len(), 1000);
+        assert_eq!(levels.logger_overrides().get("logger_1000"), None);
+    }
 
+    #[test]
+    fn log_levels_accepts_existing_key_at_capacity() {
+        let levels = LogLevels::new("info");
+        // Fill to capacity
+        for i in 0..1000 {
+            let _ = levels.set_logger_level(&format!("logger_{i}"), "debug");
+        }
 
+        // Try to update an existing key, should succeed
+        let prev = levels.set_logger_level("logger_999", "warn");
+        assert_eq!(prev.as_deref(), Some("debug"));
+        assert_eq!(levels.logger_overrides().len(), 1000);
+        assert_eq!(
+            levels
+                .logger_overrides()
+                .get("logger_999")
+                .map(String::as_str),
+            Some("warn")
+        );
+    }
 
+    #[test]
+    fn task_registry_records_multiple_successes_and_failures() {
+        let registry = TaskRegistry::new();
+        registry.register("my_task", "cron * * * * *");
 
+        // 1st success
+        registry.record_start("my_task");
+        registry.record_success("my_task", 100);
 
+        // 2nd success
+        registry.record_start("my_task");
+        registry.record_success("my_task", 110);
 
+        let snapshot = registry.snapshot();
+        let task = &snapshot["my_task"];
+        assert_eq!(task.total_runs, 2);
+        assert_eq!(task.total_failures, 0);
 
+        // 1st failure
+        registry.record_start("my_task");
+        registry.record_failure("my_task", 50, "failed");
 
+        let snapshot2 = registry.snapshot();
+        let task2 = &snapshot2["my_task"];
+        assert_eq!(task2.total_runs, 3);
+        assert_eq!(task2.total_failures, 1);
+    }
 
+    #[test]
+    fn configprops_tracks_custom_profile() {
+        let mut props = HashMap::new();
+        ConfigProperties::track_property(
+            &mut props,
+            "log.level",
+            "debug",
+            "info",
+            "custom_profile",
+        );
+        assert_eq!(props["log.level"].source, "autumn.toml");
+    }
 
+    #[test]
+    fn configprops_tracks_dev_prod_profiles() {
+        let mut props = HashMap::new();
+        ConfigProperties::track_property(&mut props, "log.level", "debug", "info", "dev");
+        assert_eq!(props["log.level"].source, "profile_default:dev");
 
+        ConfigProperties::track_property(&mut props, "log.format", "json", "text", "prod");
+        assert_eq!(props["log.format"].source, "profile_default:prod");
+    }
+
+    #[test]
+    fn configprops_returns_default_when_values_match() {
+        let mut props = HashMap::new();
+        ConfigProperties::track_property(&mut props, "log.level", "info", "info", "dev");
+        assert_eq!(props["log.level"].source, "default");
+    }
+}
+
+#[cfg(test)]
+mod havoc_proptest {
+    use super::*;
+    use proptest::prelude::*;
+
+    proptest! {
+        #![proptest_config(ProptestConfig::with_cases(1))]
+        #[test]
+        fn log_levels_memory_exhaustion(names in proptest::collection::vec(".*", 5000)) {
+            let levels = LogLevels::new("info");
+            for name in names {
+                let _ = levels.set_logger_level(&name, "debug");
+            }
+            assert!(levels.logger_overrides().len() <= 1000, "Memory leak: unbounded loggers inserted");
+        }
+    }
 }
