+++ autumn/src/auth.rs
@@ -441,9 +441,9 @@ mod tests {
             started_at: std::time::Instant::now(),
             health_detailed: false,
             metrics: crate::middleware::MetricsCollector::new(),
-            log_levels: crate::diagnostics::LogLevels::new("info"),
-            task_registry: crate::diagnostics::TaskRegistry::new(),
-            config_props: crate::diagnostics::ConfigProperties::default(),
+            log_levels: crate::actuator::LogLevels::new("info"),
+            task_registry: crate::actuator::TaskRegistry::new(),
+            config_props: crate::actuator::ConfigProperties::default(),
             #[cfg(feature = "ws")]
             channels: crate::channels::Channels::new(32),
             #[cfg(feature = "ws")]
@@ -488,9 +488,9 @@ mod tests {
             started_at: std::time::Instant::now(),
             health_detailed: false,
             metrics: crate::middleware::MetricsCollector::new(),
-            log_levels: crate::diagnostics::LogLevels::new("info"),
-            task_registry: crate::diagnostics::TaskRegistry::new(),
-            config_props: crate::diagnostics::ConfigProperties::default(),
+            log_levels: crate::actuator::LogLevels::new("info"),
+            task_registry: crate::actuator::TaskRegistry::new(),
+            config_props: crate::actuator::ConfigProperties::default(),
             #[cfg(feature = "ws")]
             channels: crate::channels::Channels::new(32),
             #[cfg(feature = "ws")]
@@ -543,9 +543,9 @@ mod tests {
             started_at: std::time::Instant::now(),
             health_detailed: false,
             metrics: crate::middleware::MetricsCollector::new(),
-            log_levels: crate::diagnostics::LogLevels::new("info"),
-            task_registry: crate::diagnostics::TaskRegistry::new(),
-            config_props: crate::diagnostics::ConfigProperties::default(),
+            log_levels: crate::actuator::LogLevels::new("info"),
+            task_registry: crate::actuator::TaskRegistry::new(),
+            config_props: crate::actuator::ConfigProperties::default(),
             #[cfg(feature = "ws")]
             channels: crate::channels::Channels::new(32),
             #[cfg(feature = "ws")]
@@ -652,9 +652,9 @@ mod tests {
             started_at: std::time::Instant::now(),
             health_detailed: false,
             metrics: crate::middleware::MetricsCollector::new(),
-            log_levels: crate::diagnostics::LogLevels::new("info"),
-            task_registry: crate::diagnostics::TaskRegistry::new(),
-            config_props: crate::diagnostics::ConfigProperties::default(),
+            log_levels: crate::actuator::LogLevels::new("info"),
+            task_registry: crate::actuator::TaskRegistry::new(),
+            config_props: crate::actuator::ConfigProperties::default(),
             #[cfg(feature = "ws")]
             channels: crate::channels::Channels::new(32),
             #[cfg(feature = "ws")]
@@ -712,9 +712,9 @@ mod tests {
             started_at: std::time::Instant::now(),
             health_detailed: false,
             metrics: crate::middleware::MetricsCollector::new(),
-            log_levels: crate::diagnostics::LogLevels::new("info"),
-            task_registry: crate::diagnostics::TaskRegistry::new(),
-            config_props: crate::diagnostics::ConfigProperties::default(),
+            log_levels: crate::actuator::LogLevels::new("info"),
+            task_registry: crate::actuator::TaskRegistry::new(),
+            config_props: crate::actuator::ConfigProperties::default(),
             #[cfg(feature = "ws")]
             channels: crate::channels::Channels::new(32),
             #[cfg(feature = "ws")]
@@ -777,9 +777,9 @@ mod tests {
             started_at: std::time::Instant::now(),
             health_detailed: false,
             metrics: crate::middleware::MetricsCollector::new(),
-            log_levels: crate::diagnostics::LogLevels::new("info"),
-            task_registry: crate::diagnostics::TaskRegistry::new(),
-            config_props: crate::diagnostics::ConfigProperties::default(),
+            log_levels: crate::actuator::LogLevels::new("info"),
+            task_registry: crate::actuator::TaskRegistry::new(),
+            config_props: crate::actuator::ConfigProperties::default(),
             #[cfg(feature = "ws")]
             channels: crate::channels::Channels::new(32),
             #[cfg(feature = "ws")]
@@ -838,9 +838,9 @@ mod tests {
             started_at: std::time::Instant::now(),
             health_detailed: false,
             metrics: crate::middleware::MetricsCollector::new(),
-            log_levels: crate::diagnostics::LogLevels::new("info"),
-            task_registry: crate::diagnostics::TaskRegistry::new(),
-            config_props: crate::diagnostics::ConfigProperties::default(),
+            log_levels: crate::actuator::LogLevels::new("info"),
+            task_registry: crate::actuator::TaskRegistry::new(),
+            config_props: crate::actuator::ConfigProperties::default(),
             #[cfg(feature = "ws")]
             channels: crate::channels::Channels::new(32),
             #[cfg(feature = "ws")]
@@ -893,9 +893,9 @@ mod tests {
             started_at: std::time::Instant::now(),
             health_detailed: false,
             metrics: crate::middleware::MetricsCollector::new(),
-            log_levels: crate::diagnostics::LogLevels::new("info"),
-            task_registry: crate::diagnostics::TaskRegistry::new(),
-            config_props: crate::diagnostics::ConfigProperties::default(),
+            log_levels: crate::actuator::LogLevels::new("info"),
+            task_registry: crate::actuator::TaskRegistry::new(),
+            config_props: crate::actuator::ConfigProperties::default(),
             #[cfg(feature = "ws")]
             channels: crate::channels::Channels::new(32),
             #[cfg(feature = "ws")]
